requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
Cython==0.29.26
dill==0.3.5.1
dm-tree==0.1.7
etils==0.6.0
flatbuffers==2.0
fonttools==4.33.3
gast==0.4.0
google-auth==2.7.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.4
grpcio==1.46.3
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.8.0
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
libclang==14.0.1
Markdown==3.3.7
matplotlib==3.5.2
numpy==1.22.0
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==8.4.0
promise==2.3
protobuf==3.20.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-addons==0.17.0
tensorflow-datasets==4.6.0
tensorflow-estimator==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.23.1
tensorflow-metadata==1.9.0
tensorflow-model-optimization==0.7.2
termcolor==1.1.0
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.64.0
typeguard==2.13.3
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
wrapt==1.14.1
zipp==3.8.0
